Fix: Supabase workspace push no longer errors on updatedAt overflow (#111)
updatedAt holds epoch-milliseconds (~1.78e12), which overflows Postgres
INTEGER (int4, max ~2.1e9). Pushing a workspace to a Supabase backend
failed with "value ... is out of range for type integer" (22003).

- schema.js sqlType(): updatedAt now maps to BIGINT (SQLite/Turso accept
  BIGINT too via integer affinity, so this is safe for every adapter).
- supabase-bootstrap.sql: CREATE columns as BIGINT + 6 idempotent
  ALTER TABLE ... TYPE BIGINT migrations for already-provisioned tables.
- tests/run.js: ratchet asserting the provisioning DDL types updatedAt as
  BIGINT and never INTEGER.
- sw.js CACHE_NAME v79 -> v80.

+++ app/sources/schema.js
@@ -95,7 +95,11 @@
   WS.TABLE_NAMES = WS.WORKSPACE_TABLES.map(function (t) { return t.name; });
 
   function sqlType(col) {
-    return col === "updatedAt" ? "INTEGER" : "TEXT";
+    // updatedAt holds epoch-MILLISECONDS (~1.78e12), which overflows Postgres
+    // INTEGER (int4, max ~2.1e9) — Supabase writes fail with "value … is out of
+    // range for type integer" (22003). BIGINT holds it; SQLite/Turso accept
+    // BIGINT too (integer affinity), so this is safe for every adapter.
+    return col === "updatedAt" ? "BIGINT" : "TEXT";
   }
 
   // A promoted column's value, normalised for a scalar SQL cell. updatedAt is
